Restructured ChessServer\GameMode
